2021.11.30 : 채용공고 삭제 - CSRF 토큰 부여 기능 추가.

+++ src/main/java/com/spring/sideproject/recruitboard/controller/RecruitBoardController.java
@@ -346,7 +346,13 @@ public class RecruitBoardController {
 	@RequestMapping("/recruitBoard/recruitBoardDelete.do/{boardId}")
 	public String doRecruitBoardDeleteAction(
 			@PathVariable int boardId
-			, @SessionAttribute(Session.USER) RecruitMemberVo recruitMemberVo) {
+			, @SessionAttribute(Session.USER) RecruitMemberVo recruitMemberVo
+			, @RequestParam String token
+			, @SessionAttribute(Session.CSRF_TOKEN) String sessionToken) {
+		
+		if ( !token.equals(sessionToken) ) {
+			throw new RuntimeException("잘못된 인증");
+		}		
 		
 		boolean isSuccess = this.recruitBoardService.deleteOneRecruitBoardByBoardIdService(boardId);
